Delete button works now (someone misunderstood Android's rm binary)
This may be an issue for various versions of rm i.e. there may be a need
to check which kind of 'rm' someone is using and adapt. The vanilla
android version of 'rm' does not understand the -f flag. To remove a
file, just say rm (no need for rm -f). To remove a directory, say
rm -r <dirname>. I don't know if the android version never needs
force, but anytime I saw the "f" passed the rm binary thought we were
trying to delete a file (or directory) called 'f', and ignored the
intended directory or file