createMountPath and cryptsetup works
Primary contribution is a modification of cryptsetup command to remove need
for manual input / output. Working with Process IO in Java is tricky at best,
and Android makes it even worse. Added the quiet flag that removed the "YES"
check, and passed in the password. We can now use the standard doShellCommand
to run cryptSetup

Also formatted the file according to java conventions