luksFormat and luksOpen now work. Removed IO parsing by changing shell commands
Passing in the key by echo'ing it onto the stdout and then telling
cryptsetup that the key was located on stdin.

Removed a number of edge conditions and inlined command parsing and reused
the ServiceShellUtils.blah command (which we know works).

Removed some dangerous flags from mke2fs in formatMOuntPath and documented that
method